added backgrounds and updated events
added backgrounds to some of the sections in the website. also updated the events section and added a fun quests button that shows the other side quests.
changed class name of a div in stats component from "stats" to "stats_new" because of clash of class name with "stats" from about component